+++ src/Dx.Domain.Primitives/SpanId.cs
@@ -14,7 +14,6 @@ using System;
 using System.Diagnostics;
 using System.Globalization;
 
-using Dx.Domain;
 
 namespace Dx.Domain.Primitives
 {